Update project, to split up code
