+++ access.log
@@ -1,0 +1,27 @@
+{"level":"info","message":"Not registering as no valid config is available","timestamp":"2016-10-27T09:23:32.424Z"}
+{"level":"info","message":"The WebSocket server is listening...","timestamp":"2016-10-27T09:23:32.434Z"}
+{"level":"debug","message":"Connection closed.","timestamp":"2016-10-27T09:23:54.378Z"}
+{"level":"debug","message":"Connection closed.","timestamp":"2016-10-27T09:27:18.019Z"}
+{"level":"info","message":"The server is exiting","timestamp":"2016-10-27T09:27:18.026Z"}
+{"level":"info","message":"Not registering as no valid config is available","timestamp":"2016-10-28T12:37:15.346Z"}
+{"level":"info","message":"The WebSocket server is listening...","timestamp":"2016-10-28T12:37:15.356Z"}
+{"level":"debug","message":"Connection closed.","timestamp":"2016-10-28T12:37:46.130Z"}
+{"level":"info","message":"The server is exiting","timestamp":"2016-10-28T12:37:50.533Z"}
+{"level":"info","message":"Not registering as no valid config is available","timestamp":"2016-10-28T12:44:21.349Z"}
+{"level":"info","message":"The WebSocket server is listening...","timestamp":"2016-10-28T12:44:21.359Z"}
+{"level":"info","message":"The server is exiting","timestamp":"2016-10-28T12:44:24.530Z"}
+{"level":"info","message":"Not registering as no valid config is available","timestamp":"2016-10-28T12:51:12.875Z"}
+{"level":"info","message":"The WebSocket server is listening...","timestamp":"2016-10-28T12:51:12.886Z"}
+{"level":"info","message":"The server is exiting","timestamp":"2016-10-28T12:51:14.687Z"}
+{"level":"info","message":"Not registering as no valid config is available","timestamp":"2016-10-28T12:55:54.964Z"}
+{"level":"info","message":"The WebSocket server is listening...","timestamp":"2016-10-28T12:55:54.974Z"}
+{"level":"info","message":"The server is exiting","timestamp":"2016-10-28T12:56:00.669Z"}
+{"level":"info","message":"Not registering as no valid config is available","timestamp":"2016-10-28T12:58:13.184Z"}
+{"level":"info","message":"The WebSocket server is listening...","timestamp":"2016-10-28T12:58:13.194Z"}
+{"level":"info","message":"The server is exiting","timestamp":"2016-10-28T12:58:18.364Z"}
+{"level":"info","message":"Not registering as no valid config is available","timestamp":"2016-10-28T13:07:35.184Z"}
+{"level":"info","message":"The WebSocket server is listening...","timestamp":"2016-10-28T13:07:35.194Z"}
+{"level":"info","message":"The server is exiting","timestamp":"2016-10-28T13:07:37.193Z"}
+{"level":"info","message":"Not registering as no valid config is available","timestamp":"2016-10-28T13:09:42.909Z"}
+{"level":"info","message":"The WebSocket server is listening...","timestamp":"2016-10-28T13:09:42.919Z"}
+{"level":"info","message":"The server is exiting","timestamp":"2016-10-28T13:09:49.720Z"}
