+++ manifest.yml
@@ -1,0 +1,8 @@
+applications:
+- path: .
+  memory: 256M
+  instances: 1
+  domain: mybluemix.net 
+  name: kilian-node-room
+  host: kilian-node-room
+  disk_quota: 1024M
