+++ src/ws/sendExamine.js
@@ -1,0 +1,22 @@
+var sendExamine = function(conn, target, username, logger)
+{
+  logger.debug("Target \"" + target + "\" asked for examination.")
+  var sendTarget = target
+  var sendMessageType = "player"
+  var messageObject = {
+    type: "event",
+    bookmark: 2223,
+    content: {
+    }
+  }
+
+  messageObject.content[target] = "We having nothing in here to really examine."
+
+  var messageToSend = sendMessageType + "," +
+            sendTarget + "," +
+            JSON.stringify(messageObject)
+
+  conn.sendText(messageToSend)
+}
+
+module.exports = sendExamine;
