+++ src/ws/sendUnknownCommand.js
@@ -1,0 +1,20 @@
+var sendUnknownCommand = function(conn, target, content, logger) {
+  logger.info("Unknown command from user: " + content)
+  var responseObject = {
+          type: "event",
+          "content": {
+          }
+        }
+
+    responseObject.content[target] = "Node.js looked at your command, and barfed."
+        var sendMessageType = "player"
+        var sendTarget = target
+
+        var messageText = sendMessageType + "," +
+                  sendTarget + "," +
+                  JSON.stringify(responseObject)
+
+    conn.sendText(messageText)
+}
+
+module.exports = sendUnknownCommand;
